[FIX] docker-compose:  Use correct volumes
Standalone needs a named volume to keep data saved even if the container is deleted.

Hosted is the only one needing to map modules from the computer, standalone uses modules inside the container.

SSH keys are mapped for both implementations.

This also allows a faster execution for standalone, which is the recommended approach from now on.